ScatterPlotViewer: split DataSeriesControl into own component
