+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/ScatterPlotViewer/ScatterPlotViewer.jsx
@@ -3,37 +3,10 @@ import PropTypes from 'prop-types'
 import { ParentSize } from '@visx/responsive'
 import ZoomingScatterPlot from './components/ZoomingScatterPlot'
 import ScatterPlot from './components/ScatterPlot'
+import DataSeriesControls from './components/DataSeriesControls'
 import ZoomControlButton from '../ZoomControlButton'
 
 const DEFAULT_HANDLER = () => {}
-
-function DataSeriesControls ({
-  fullData,
-  indexesToHide = [],
-  toggleIndex = DEFAULT_HANDLER
-}) {
-  if (!fullData) return null
-
-  function onCheckboxChange (event) {
-    const index = parseInt(event?.currentTarget?.dataset?.index)
-    if (!Number.isNaN(index)) toggleIndex(index)
-  }
-
-  return (
-    <ul className='ScatterPlotViewer-DataSeriesControls'>
-      {fullData?.map((dataSeries, index) => (
-        <li key={`data-series-${index}`}>
-          <input type='checkbox'
-            data-index={index}
-            checked={!indexesToHide.includes(index)}
-            onChange={onCheckboxChange}
-          />
-          <label>{dataSeries.seriesOptions?.label?.trim() || `series ${index+1}`}</label>
-        </li>
-      ))}
-    </ul>
-  )
-}
 
 const ScatterPlotViewer = forwardRef(function ScatterPlotViewer (props, ref) {
   const {
