+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/ScatterPlotViewer/components/DataSeriesControls/index.js
@@ -1,0 +1,1 @@
+export { default } from './DataSeriesControls'
